+++ app/src/test/java/nodomain/freeyourgadget/gadgetbridge/test/CalendarEventTest.java
@@ -1,5 +1,6 @@
 package nodomain.freeyourgadget.gadgetbridge.test;
 
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.ArrayList;